Implemented everything in single precision

+++ larank/decision_function.h
@@ -34,20 +34,20 @@ public:
 
     virtual ~DecisionFunction ();
 
-    double computeGradient (const Eigen::VectorXf &features, int label, int this_label) const;
-    double computeScore (const Eigen::VectorXf &features) const;
+    float computeGradient (const Eigen::VectorXf &features, int label, int this_label) const;
+    float computeScore (const Eigen::VectorXf &features) const;
 
-    void update (const Eigen::VectorXf &features, double lambda, int64_t pattern_id);
+    void update (const Eigen::VectorXf &features, float lambda, int64_t pattern_id);
 
-    double getBeta (int64_t pattern_id) const;
+    float getBeta (int64_t pattern_id) const;
     bool isSupportVector (int64_t pattern_id) const;
 
     int getNSV () const;
-    double getW2 () const;
+    float getW2 () const;
 
 private:
     // Beta (indicator) values of each support vector
-    std::unordered_map<int64_t, double> beta;
+    std::unordered_map<int64_t, float> beta;
 
     // Hyperplane weights
     Eigen::VectorXf w;
